+++ ndc-lambda-sdk/src/host.ts
@@ -1,3 +1,7 @@
+// We must initialize OpenTelemetry instrumentation before importing any other module
+import * as sdkInstrumentation from "@hasura/ndc-sdk-typescript/instrumentation";
+sdkInstrumentation.initTelemetry("ndc-lambda-sdk");
+
 import * as sdk from "@hasura/ndc-sdk-typescript";
 import { createConnector } from "./connector";
 import { makeCommand } from "./cmdline";